Moar Power!! Fight Pirates!!
The ranger-devs are happy with the modifications to the attack vehicles. They need one last modification, though, so that they can know what sort of firepower they are able to use on each vehicle.

Add a property called # of weapons to each vehicle using appropriate notation. Assign these new properties the following values, inside the appropriate object:

There are 8 torpedoes.
There are 4 lasers.
There is 1 rear-mounted slingshot.
The original objects and previously made modifications are provided for you.