Blinders On!
The ranger-devs also want to upgrade Lighthouse Rock with some super-blinding light bulbs, while sending the existing bulbs to be used at the Underwater Outpost. Below is the Lighthouse Rock object, and an array of super-blinders. Each index of the superblinder array contains a bulb name and its wattage within a subarray. Do the following:

Completely remove the existing bulb array from the Lighthouse Rock object, including its property.
Add in the new array using only its variable name and the property name weaponBulbs.
Examine the array and write a line of code that logs out the name of the superblinder with the highest wattage using only object and array notation (no hard-coding the message with the name).