Man your Bulb Stations, Doods!
This one’s a little tricky…think carefully. Now that we’ve got some man- and woman-power at the Lighthouse, we need to tell the ranger-devs which super-blinding bulb they are stationed at.

Build a function that takes in a location object, examines every ranger’s station number, and retrieves the name of the super-blinder to which they’re assigned from the array of the super-blinders. When the right super-blinder is found for a ranger-dev, alert it in a message to the ranger in the following format:

Avast, me hearties!
There be Pirates nearby! Stations!
<name>, man the <superblinder>!
<name>, man the <superblinder>!
<name>, man the <superblinder>!
In other words, the entire list of assignments should be contained within one alert. Hmm, sound familiar?

Call your function dontPanic, and when it is completed, call it on the lighthouseRock variable which is provided for you, complete with new ranger-dev objects as if they had been there from the get-go.