Functions As Properties II Hmm, looks like Jordan, the new ranger at Lighthouse Rock, doesn’t have a station because there aren’t enough super-blinding bulbs to go around! What’s a dual-wielding hand crossbow master to do?
The ranger-devs ask you to build a new function property for lighthouseRock that adds a new super-blinder bulb to their list, complete with name and insane wattage. This function property, however, needs to be added on the fly (they are picky about strange things).

Name your new property addBulb, and have it take parameters name and wattage. Remember, also, that you will be adding this property to the lighthouseRock object, which is provided for you in the editor with some slight reorganization.

The current super-blinder array is below. Be careful to notice how the data is arranged and indexed. Add any new bulb created to the back of the array. Additionally, you’ll need to add the bulb by calling the lighthouseRock object’s property that REFERS to the array, rather than the external array.

var superBlinders = [ ["Firestorm", 4000],
                      ["Solar Death Ray", 6000],
                      ["Supernova", 12000] ];