Enumeration I
The devs at Lighthouse Rock have received a shipment of Spearguns for use in the battle against the Really Awesome Pirates (RAP). They need to know quickly on their augmented reality eyeglasses just which spears have arrived (they can’t leave their posts at the bulbs). The Object for the shipping box is below:

var rockSpearguns = {
  Sharpshooter: {barbs: 2, weight: 10, heft: "overhand"},
  Pokepistol: {barbs: 4, weight: 8, heft: "shoulder"},
  Javelinjet: {barbs: 4, weight: 12, heft: "waist"},
  Firefork: {barbs: 6, weight: 8, heft: "overhand"},
  "The Impaler": {barbs: 1, weight: 30, heft: "chest"}
};
Build a declared function called listGuns that accepts a container of spearguns, called guns, as a parameter and logs out the name of each Speargun. Additionally, use speargun to refer to the property names inside your function. When your function is built, call it on the given shipment of guns.