Prototypes I
The cowboy-devs at the Plains of Prototypes seek your assistance. They’ve got a big herd-merger to manage, and they have multiple lists of cattle being imported from multiple regions. Each individual bovine is represented as a JavaScript object, and multiple bovine objects are contained together as an unordered list within an array. Here’s an example, which happens to be the list that came over from good old Conditional Canyon:

var canyonCows = [
  {name: "Bessie", type: "cow", hadCalf: "Burt"},
  {name: "Donald", type: "bull", hadCalf: null},
  {name: "Esther", type: "calf", hadCalf: null},
  {name: "Burt", type: "calf", hadCalf: null},
  {name: "Sarah", type: "cow", hadCalf: "Esther"},
  {name: "Samson", type: "bull", hadCalf: null},
  {name: "Delilah", type: "cow", hadCalf: null}
];
The only problem is that the cowboy-devs don’t have time to go through the dozens of lists they are getting and count each specific type of cattle being imported, i.e. the number of cows, bulls, or a calfs. Since the only use the cowboys have for arrays is to make these lists of cattle objects, they’ve asked you to go ahead and add a function, called countCattle, to the Array prototype that will allow any array list of Cows to access precisely how many bulls, cows, or calfs are present in the array. The function should accept a parameter and return a number value.