Prototypes III
To let current mother cows rest, the cowboy-devs at the Plains of Prototypes have decided only to breed those cows that have not given birth to any calves yet. Since the only use the cowboy-devs have for the JavaScript object is in making cattle representations (it’s a rough life), they want to go ahead and add some functionality to the Object prototype itself. They’ve asked you to build two functions in order, explained here:

A function called noCalvesYet, which will be added to the Object prototype, and will return true for an object if the object is a cow with no calves…and false otherwise.

A function called countForBreeding, which will be added to the Array prototype, and will examine an array’s list of cattle objects, counting all cows that are not moms. To do this, you should employ the function in criteria number (1), as well as use a counter variable called numToBreed. Ultimately, this function should return numToBreed representing precisely how many non-mother cows are available within any array.

A list of cattle is provided below for your reference:

var forestCows = [
  {name: "Legolas", type: "calf", hadCalf: null},
  {name: "Gimli", type: "bull", hadCalf: null},
  {name: "Arwen", type: "cow", hadCalf: null},
  {name: "Galadriel", type: "cow", hadCalf: null},
  {name: "Eowyn", type: "cow", hadCalf: "Legolas"}
];