Creation with Prototypes I
The cowboy-devs at the Plains of Prototypes admire your excellent work with their landscape’s namesake. They have a new problem for you to examine, one that they are really excited about. For the first time, they have a new robotic fence post system for their pastures. This system is pretty cool, because it allows immediate shifting of the wheel-based fence posts to make a forced migration of the cows into different, healthier pastures for grazing.

Here’s the generic fence post object, starting at coordinates (0,0), which is the location of the main ranch:

var genericPost =
  {x: 0, y: 0, postNum: undefined,
   connectionsTo: undefined,
   sendRopeTo: function ( connectedPost ) {
     if(this.connectionsTo == undefined){
     var postArray = [ ];
     postArray.push(connectedPost);
       this.connectionsTo = postArray;
     } else {
       this.connectionsTo.push(connectedPost);
     }
   }
  };
Each new fence post will travel, on its own, out to the coordinates assigned to it when a new object for it is generated. There it will send out robotically guided connection ropes to other, appropriately designated fence posts to enclose a pasture. Bear in mind that it takes two ropes to for fence posts to be connected; i.e. each sends a rope to the other. Examine the prototype closely to see how the system is designed to take care of that.

To test the creation of these new fence posts objects, the cowboy-devs have a few requirements:

Use the Object.create method and the base fence post as a prototype to create two new fence posts, post1 and post2.
Send the first fence post to (-2, 4) using standard property modification with dots only. Also set its number to 1.
Send the second fence post to (5, 1) using standard property modification with dots only. Also set its number to 2.
Use the method inherited from the prototype to connect the fence posts together. Each connectionsTo array should receive an entire object.
Build some posts, (wo)man!