Constructor I
So now the cowboy-devs are so excited about their robotic fence posts that they’ve decided to go ahead and make a constructor function for fence posts, to which they can pass a few parameters and get back a new instance of a post.

And, of course, by making a new constructor function, they mean YOU will make it.

Your job is to take your existing ‘genericPost’ and rebuild all of its properties inside a single Fencepost constructor (we’ll worry about inheritance in the next challenge). The ‘genericPost’ is provided for you:

var genericPost =
  {x: 0, y: 0, postNum: undefined,
   connectionsTo: undefined,
   sendRopeTo: function ( connectedPost ) {
     if(this.connectionsTo == undefined){
       var postArray = [ ];
       postArray.push(connectedPost);
       this.connectionsTo = postArray;
     } else {
       this.connectionsTo.push(connectedPost);
     }
   }
  };
Your constructor should take in parameters with names that shadow the property names in genericPost, with one exception. The cowboy-devs have asked you to set the connectionsTo property initially to be an empty array inside the constructor (no parameter needed). Then, modify the sendRopeTo function in a way that EXPECTS an array to already be present, rather than having to make one as it currently does when it finds an undefined value. This, by the way, will make the function much smaller.

Remember, for now, you’ll also build any functions as properties directly within the constructor.