Constructor III
So now that’s there’s eleventy-billion fence posts everywhere, the cowboy-devs have noticed a significant drain on their memory resources. They’d like you to take a look around the Fencepost constructor and see if there’s anything you can add to a prototype, so that every stinkin’ fence post doesn’t have to carry around anything that it could get from just one place.

Below is the current status of the constructor, with some additions the cowboy-devs have made to improve functionality of the fence post objects. Your job is to identify the portions of the constructor that should be available to ALL fenceposts, and put those in a prototype for fence posts. Your answer should include the modified constructor as well as the newly designed prototype for that constructor. Good luck, pardner.

function Fencepost (x, y, postNum){
  this.x = x;
  this.y = y;
  this.postNum = postNum;
  this.connectionsTo = [];
  this.sendRopeTo = function ( connectedPost ){
    this.connectionsTo.push(connectedPost);
  };
  this.removeRope = function ( removeTo ){
  var temp = [];
  for(var i = 0; i<this.connectionsTo.length; i++){
     if(this.connectionsTo[i].postNum != removeTo){
       temp.push(this.connectionsTo[i]);
     }
  }
    this.connectionsTo = temp;
  }
  this.movePost = function (x, y){
    this.x = x;
    this.y = y;
  };
}