Overriding Prototypal Methods II
Next up, the Plains of Prototypes cowboy-devs have asked for an improved toString method for their fence posts. An example of the string message they’d like is below:

Fence post #10:
Connected to posts:
11
12
13
Distance from ranch: 5 yards
Override the toString method by using the Fencepost prototype and a newly built function that returns a string message in the above format, but with unique data for each post. The constructor and all of your previous prototype additions are provided for you as a reference for property names and procedures. Remember that connectionsTo is a list of Fencepost objects.